requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.57.0
accelerate>=0.26.0
datasets>=2.20.0
Pillow>=10.0.0
tqdm
